requirements.txt
Django==5.2
pillow==11.1.0
environs==14.1.1
pillow==11.1.0 
aiogram==3.19.0
psycopg2==2.9.10
python-telegram-bot==13.15
python-dotenv==1.1.0
asgiref==3.8.1
sqlparse==0.5.3
tzlocal==5.3.1